requirements.txt
asgiref==3.5.2
Django==4.0.4
django-extensions==3.1.5
Pillow==9.1.1
sqlparse==0.4.2
tzdata==2022.1
djangorestframework~=3.13.1